New component for value field
